code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
a=1.0=pypi_0
absl-py=1.4.0=pypi_0
albumentations=0.4.3=pypi_0
altair=4.2.2=pypi_0
attrs=23.1.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blas=1.0=mkl
blinker=1.6.2=pypi_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
dataclasses=0.6=pypi_0
decorator=5.1.1=pypi_0
einops=0.3.0=pypi_0
filelock=3.12.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.4.0=pypi_0
future=0.18.3=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.17.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.0=pypi_0
idna=3.4=pypi_0
imageio=2.9.0=pypi_0
imageio-ffmpeg=0.4.2=pypi_0
imgaug=0.2.6=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=pypi_0
lazy-loader=0.2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
more-itertools=9.1.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.0.0=pypi_0
opencv-python=4.1.2.30=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openssl=1.1.1t=h7f8727e_0
pandas=1.5.3=pypi_0
pillow=9.4.0=py38h6a678d5_0
pip=20.3.3=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
protobuf=3.20.3=pypi_0
pudb=2019.2=pypi_0
pyarrow=11.0.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydeck=0.8.1b0=pypi_0
pympler=1.0.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.8.5=h7579374_1
pytorch=1.7.0=py3.8_cuda10.2.89_cudnn7.6.5_0
pytorch-lightning=1.0.8=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2023.3.23=pypi_0
requests=2.28.2=pypi_0
rich=13.3.4=pypi_0
sacremoses=0.0.53=pypi_0
scikit-image=0.20.0=pypi_0
scipy=1.9.1=pypi_0
setuptools=66.0.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.41.2=h5eee18b_0
streamlit=1.21.0=pypi_0
taming-transformers=0.0.1=dev_0
tensorboard=2.12.2=pypi_0
tensorboard-data-server=0.7.0=pypi_0
test-tube=0.7.5=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torchvision=0.8.1=py38_cu102
tqdm=4.65.0=pypi_0
transformers=4.3.1=pypi_0
typing_extensions=4.5.0=py38h06a4308_0
tzdata=2023.3=pypi_0
tzlocal=4.3=pypi_0
urllib3=1.26.15=pypi_0
urwid=2.1.2=pypi_0
validators=0.20.0=pypi_0
watchdog=3.0.0=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
